flogging_macros

Macro info

info!() { /* proc-macro */ }

Log an INFO message.

INFO is a message level for informational messages.

Typically INFO messages will be written to the console or its equivalent. So the INFO level should only be used for reasonably significant messages that will make sense to end users and system administrators.

[default level] : This is the default level used when a logger is created. See set_level! for changing this.

If the logger is currently enabled for the INFO message level then the given message is forwarded to all the registered output Handler objects.

Parameters

msg - See Special Note

Examples

See config. The syntax/usage is the same. Just substitute info! for config!.